+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.6.0 - 2024-10-07
+
+### Added
+
+- Allow setting RefUID on a per row basis.
+
 ## [0.5.0] - 2024-06-06
 
 ### Added